+++ app/lizard-nxt-master-controller.js
@@ -27,6 +27,7 @@ angular.module('lizard-nxt')
    'MapService',
    'rmAllButLastAssetAndGeometry',
    'UrlService',
+   'AssetService',
 
   function ($scope,
             $rootScope,
@@ -39,7 +40,8 @@ angular.module('lizard-nxt')
             State,
             MapService,
             rmAllButLastAssetAndGeometry,
-            UrlService) {
+            UrlService,
+            AssetService) {
 
   $scope.version = version;
   $scope.tooltips = CabinetService.createTooltips();
@@ -65,10 +67,13 @@ angular.module('lizard-nxt')
   $scope.transitionToContext = function (context) {
     if (context !== State.context) {
       State.context = context;
-      if (State.context === 'map' && State.box.type === 'point') {
-        rmAllButLastAssetAndGeometry();
+      if (context === 'map' && AssetService.getNonNestedAssets().length > 1) {
+        // If more than one asset became selected when user was in db context,
+        // we switch the box-type/tool to 'multi-point':
+        State.box.type = 'multi-point';
       }
     }
+
     var overlay = angular.element('#context-transition-overlay')[0];
     overlay.style.transition = null;
     overlay.style.minHeight = window.innerHeight + 'px';
@@ -116,7 +121,7 @@ angular.module('lizard-nxt')
     if ($window.location.host === 'portal.ddsc.nl') {
       application = 'DDSC';
     }
-    var portal = $window.location.host.split('.')[0]
+    var portal = $window.location.host.split('.')[0];
     portal = portal.charAt(0).toUpperCase() + portal.slice(1);
     return application + ' ' + portal;
   };